Remove non existing SPI NAND/FLASH from board file
When booting linux - for some reason it would think that there was an /dev/mtd0blk device
even though we only have the SD card. In DeepEddy this seems to have caused an issue when the
SPI_3.3V_!5V bit was set HIGH.

In this commit I removed all definitions of this SPI flash so it no longer shows on bootup.