Skip the firmware check if it doesn't exist
In the new upgrade model, it's possible that the firmware file doesn't
exist.  This is okay because we may be being updated using an old script
that doesn't understand the new format.  We will go ahead and be good
with the file completely missing because the client will try again to update
when the script is good.