SMART-8479: Format the root partion with ext3
Use the ext3 filesystem for the root partition because it
has better resiliency under adverse conditions.  This patch
modifies the script and initramfs used in the ug-uImage to
upgrade the system so that it contains the mke2fs from e2fsprogs because the busybox mke2fs will not format ext3, and uses this in the script to format the partition properly.

Note that this commit is a companion to commits in other
repositories to complete the changes.